Move non-template functions into source file
